+++ core/src/main/kotlin/com/bumble/appyx/core/children/ChildEntryExt.kt
@@ -5,6 +5,6 @@ import com.bumble.appyx.core.node.Node
 val <T> ChildEntry<T>.nodeOrNull: Node?
     get() =
         when (this) {
-            is ChildEntry.Eager -> node
-            is ChildEntry.Lazy -> null
+            is ChildEntry.Initialized -> node
+            is ChildEntry.Suspended -> null
         }