+++ core/src/main/kotlin/com/bumble/appyx/core/navigation/onscreen/OnScreenStateResolver.kt
@@ -1,6 +1,6 @@
 package com.bumble.appyx.core.navigation.onscreen
 
-interface OnScreenStateResolver<State> {
+fun interface OnScreenStateResolver<State> {
 
     fun isOnScreen(state: State): Boolean
 }